move gcl to legacy.
